+++ src/components/features/authForm/FooterForm.tsx
@@ -1,0 +1,32 @@
+'use client';
+
+import Image from 'next/image';
+
+const listImage = [
+  { src: '/logo_google.svg', width: 40, height: 40, alt: 'logo google' },
+  { src: '/logos_facebook.svg', width: 40, height: 40, alt: 'logo google' },
+  { src: '/logo_apple.svg', width: 40, height: 40, alt: 'logo google' },
+];
+const FooterForm = () => {
+  return (
+    <>
+      <div className="before:content-[''] before:mr-2 after:content-[''] after:ml-2 before:w-[20%] before:h-[1px] before:block flex after:h-[1px] after:w-[20%] after:block before:bg-[#373535] after:bg-[#373535] items-center justify-center">
+        <span className="m-6 text-xl">Login with</span>
+      </div>
+      <div className="flex gap-14 items-center justify-center pb-4">
+        {listImage.map((item) => (
+          <Image
+            key={item.src}
+            src={item.src}
+            width={item.width}
+            height={item.height}
+            alt={item.alt}
+            className="hover:cursor-pointer"
+          />
+        ))}
+      </div>
+    </>
+  );
+};
+
+export default FooterForm;
