+++ src/styles/globals.css
@@ -66,26 +66,6 @@
   }
 }
 
-*::-webkit-scrollbar {
-  width: 16px;
-}
-
-*::-webkit-scrollbar-track {
-  border-radius: 8px;
-}
-
-*::-webkit-scrollbar-thumb {
-  height: 56px;
-  border-radius: 8px;
-  border: 4px solid transparent;
-  background-clip: content-box;
-  background-color: #888;
-}
-
-*::-webkit-scrollbar-thumb:hover {
-  background-color: #555;
-}
-
 @layer base {
   * {
     @apply border-border;
